movielens-backend/requirements.txt
# Python dependencies for the backend API
# Web Framework
fastapi>=0.100.0,<0.112.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<2.4.0 # For loading settings from env/.env

# ASGI Server & Process Manager
uvicorn[standard]>=0.22.0,<0.30.0 # Includes uvloop, httptools for performance
gunicorn>=20.1.0,<22.1.0         # Production process manager

# Authentication & Security (JWT)
python-jose[cryptography]>=3.3.0,<3.4.0

# Database (Async MongoDB Driver)
motor>=3.1.0,<3.5.0

# Cache (Async Redis Driver)
redis>=4.5.0,<5.1.0

# Environment variable loading (useful for local dev, optional in container)
python-dotenv>=1.0.0,<1.1.0

# Structured Logging (optional but recommended for Cloud Logging)
python-json-logger>=2.0.0,<2.1.0

# Google Cloud Client Libraries (optional, if directly interacting with GCP APIs like Secret Manager)
# google-cloud-secret-manager>=2.16.0,<2.20.0

# Data Science libraries (if any heavy lifting done in API - unlikely/avoid)
# numpy>=1.21.0,<1.27.0
# scipy>=1.8.0,<1.14.0

# CORS Headers (FastAPI dependency)
python-multipart>=0.0.5,<0.0.10 # Often needed for form data, included for completeness